services/publisher_service/requirements.txt
celery[redis]
sqlalchemy
mysqlclient
python-dotenv
python-json-logger
prometheus-client
fastapi
uvicorn[standard]
python-wordpress-xmlrpc